include primary key in insert_all conflict target if specified in unique_by
As it is, when the primary key is passed in to `insert_all` as `:unique_by`, that column (or columns) are omitted from the conflict target, resulting in the conflict target being implicitly _all_ unique indexes/constraints on the table. However, in a database like PostgreSQL, some of the constraints considered may be deferred, which are not able to be considered in the context of a conflict target. In such cases, it is helpful to be able to explicitly enumerate the primary key(s) in the conflict target so that _only_ that key is considered when determining the `ON CONFLICT` behavior.