Do not move location of key generation logic
In the prior commit the location of the key and version generation was moved to a custom object. While faster, it was more confusing to understand if the logic was the same or different.

This PR works by treating the Cache::Key as a value store that cache's the value of cache_key and cache_version so that it only needs to be generated once instead of multiple times.