Add citext support for associations
The database i'm working with has an association that has to be linked with a citext attribute (it is on a table that is used to integrate with another database). While requesting the parent object with includes, i was getting "undefined method `first' for nil:NilClass", that was traced to "activerecord (6.0.0) lib/active_record/associations/preloader/association.rb:105:in `block in records_for'".

I debugged that function using the web console and found out that even if citext could detect case insensitive, it would return nil on rails because it is case sensitive. (eg: sql can relate "123F" with "123f" using citext). So, this PR proposition is that, at the convert_keys function, both keys would be upcased if either types are citext.