Add methods to the public API of ActiveRecord::Inheritance
Those methods make possible to extend STI and Polymorphic associations.

They are useful for cases where you renamed a class and the name of the
class doesn't match the data in the database.

You can now implement those methods in your model to load records with
name of classes that don't exist anymore.

A simple implementation would look like:

    class Animal < ActiveRecord::Base
      @@old_names = {
        "Lion" => "BigCat"
      }
      def self.sti_name
        name = super
        @@old_names[name] || name
      end

      def self.sti_class_for(type_name)
        @@old_names.inverse[type_name]&.constantize || super
      end
    end