Avoid shelling out for generator `generate` action
This change addresses a few issues:

First, shelling out to the `rails` command requires the destination
directory to contain certain files that the command uses to initialize
itself.  While this is not an issue when running generators normally, it
is troublesome when testing generator-invoking generators which output
to ephemeral destination directories.

Second, shelling out to the `rails` command is very slow.  This also is
not a particular concern when running generators normally, but it makes
test suites for generator-invoking generators painfully slow.

Third, shelling out to the `rails` command fails silently by default.
Such silent failures can be surprising, and can lead to confusing
downstream failures.