Make AS::Deprecation.silence block thread-local
Previouly `ActiveSupport::Deprecation.silence { ... }` would silence
deprecations from all threads, which I think is undesirable in a
mutli-threaded environment.

This leaves `ActiveSupport::Deprecation.silenced=` as setting a global
option, which might be a little confusing, but I think that these are
used in different ways, and this best preserves how they are generally
used. `silence {}` is used to silence a small section of code,
`silenced=` is usually used to set a default in an initializer or test
helper.

I somewhat think that `ActiveSupport::Deprecation.silenced=` should be
deprecated in favour of `ActiveSupport::Deprecation.behavior = :silence`