Add first & last IDs to collection cache versions
In collections with a limit, including the ID of the first and last
record when generating a collection's cache version ensures correct
cache key expiration after a record is deleted. Prior to this commit,
collection cache versions only include the collection's size and maximum
updated_at timestamp, e.g.:

    "#{collection_size}-#{max_updated_at_timestamp}"

To demonstrate why this isn't robust enough, consider the following
database table:

    Employee.all

    SELECT * FROM employees;
    +----+-------+---------------------+---------------------+
    | id | name  | created_at          | updated_at          |
    +----+-------+---------------------+---------------------+
    |  1 | alice | 2019-11-01 00:00:00 | 2019-11-01 00:00:00 |
    |  2 | carol | 2019-11-02 00:00:00 | 2019-11-02 00:00:00 |
    |  3 | doug  | 2019-11-03 00:00:00 | 2019-11-03 00:00:00 |
    |  4 | bob   | 2019-11-04 00:00:00 | 2019-11-04 00:00:00 |
    +----+-------+---------------------+---------------------+

Select the first three employees sorted alphabetically:

    employees = Employee.order(:name).limit(3)

    SELECT * FROM employees ORDER BY name ASC LIMIT 3;
    +----+-------+---------------------+---------------------+
    | id | name  | created_at          | updated_at          |
    +----+-------+---------------------+---------------------+
    |  1 | alice | 2019-11-01 00:00:00 | 2019-11-01 00:00:00 |
    |  4 | bob   | 2019-11-04 00:00:00 | 2019-11-04 00:00:00 |
    |  2 | carol | 2019-11-02 00:00:00 | 2019-11-02 00:00:00 |
    +----+-------+---------------------+---------------------+

Note the collection's size (3) and max updated_at timestamp
(2019-11-04 00:00:00), both of which are used to compose the cache
version:

    employees.cache_version
    # => "3-20191104000000"

Now, delete an employee:

    Employee.find_by(name: "alice").destroy

    DELETE FROM employees WHERE name = "alice";

Re-run the original query:

    employees = Employee.order(:name).limit(3)

    SELECT * FROM employees ORDER BY name ASC LIMIT 3;
    +----+-------+---------------------+---------------------+
    | id | name  | created_at          | updated_at          |
    +----+-------+---------------------+---------------------+
    |  4 | bob   | 2019-11-04 00:00:00 | 2019-11-04 00:00:00 |
    |  2 | carol | 2019-11-02 00:00:00 | 2019-11-02 00:00:00 |
    |  3 | doug  | 2019-11-03 00:00:00 | 2019-11-03 00:00:00 |
    +----+-------+---------------------+---------------------+

Note that while the contents of the collection has changed, its size and
maximum updated_at timestamps remain the same, resulting in the
cache_version failing to correctly expire:

    employees.cache_version
    # => "3-20191104000000"

Including the ID of the first and last records in a collection fixes
this problem - when deleting a record from a collection with a limit,
either the collection size will shrink or one of its bounding IDs will
change.

If a collection is loaded, determining the IDs of its first and last
record is straightforward:

    employees = Employee.limit(3)
    first_id = employees.size > 0 ? employees.first.id : nil
    last_id = employees.size > 0 ? employees.last.id : nil

If a collection isn't loaded, this commit uses SQL common table
expressions (for SQLite and PostgreSQL databases) or user-defined
variables (for MySQL databases) to determine the collection's first and
last IDs. Example SQL queries below (table and column aliases shortened
for readability):

    -- PostgreSQL/SQLite
    WITH
      col AS
        (SELECT id, updated_at FROM users ORDER BY name ASC LIMIT 3),
      agg AS
        (SELECT COUNT(*) AS size, MAX(updated_at) AS timestamp FROM col)
    SELECT
      size,
      timestamp,
      (SELECT id FROM col LIMIT 1) AS first_id,
      (SELECT id FROM col LIMIT 1 OFFSET
        GREATEST((SELECT size FROM agg) - 1, 0)) AS last_id
    FROM agg;

    -- MySQL
    SET @first_id := NULL, @last_id := NULL;
    SELECT
      COUNT(*) AS size,
      MAX(updated_at) AS timestamp,
      @first_id AS first_id,
      @last_id AS last_id
    FROM (
      SELECT
        @first_id := COALESCE(@first_id, id),
        @last_id := id,
        updated_at
      FROM users
      ORDER BY name ASC LIMIT 3
    ) col;

    +------+----------+---------+---------------------+
    | size | first_id | last_id | timestamp           |
    +------+----------+---------+---------------------+
    |    3 |        4 |       3 | 2019-11-04 00:00:00 |
    +------+----------+---------+---------------------+

Additionally, this commit updates the signatures of
ActiveRecord::Relation#cache_version, #cache_key and related methods to
use keyword instead of positional arguments, as said methods now accept
multiple optional arguments. In order to avoid breaking changes, both
methods still accept positional arguments, but will raise deprecation
warnings if passed.

Fixes #31996, #34408 and #37555.