Delegate whether or not the QC is enabled to the pool
This patch delegates whether or not the QC is enabled to the pool, but
maintains the ability to configure query cache at the individual
connection level.

I want to do this because we are currently storing the state of the QC
in two places: the pool *and* the connection.  That means we have to
make sure the connection and pool are in sync with each other.  This is
frustrating for debugging, but also means we have to do extra work when
checking out the connection.

I'd like for there to be one single source of truth as to the state of
the QC.  This commit means we have a single source, unless a user
specifically overrides the value.