Clarify documentation on ActiveRecord Callbacks
Mainly these changes aim to make this documentation a bit more ESL (English
as a second language friendly).

Here are some examples of the things I changed:
- replace the constant use of `get <verb>` with `is <verb>`
- replace $10 words like `alteration` with clearer $1 words like `change`
- `give you immense power` -> `offer you a lot of control`
- missing `<code>` wrapping to denote method/variable references and SQL
statements like `COMMIT` and `ROLLBACK`.
- clarify the ordering callback example specifically to explain what children is
(an association with records)

This is minimal stuff but I think it improves the clarity and legibility of this
extremely important piece of documentation. I'm actually surprised that it's not
linked to from the [Callbacks Guide][1]

[1]: https://guides.rubyonrails.org/active_record_callbacks.html