Simplify example of Action Mailbox usage [CI SKIP]
The example in the documentation looked like it had been extracted from
an actual application, with some simplifications made. However, it still
contained a fair amount of application specific logic that wasn't
relevant to demonstrating use of ActionMailbox::Base.

I simplified the example to limit it to just demonstrating ActionMailbox
usage, and changed the naming of things like `forwarder` to `user` to
make it clearer it is application specific.