Refactor generator `route` action tests
Add `assert_routes` helper, which verifies that routes are always added
to the `Rails.application.routes.draw` block, are always indented, and
are always terminated with a newline.  This "fixes" one test that was
not testing something it claimed to, and obsoletes another test.

Also, add a test case for indentation of multi-line routing code.