update getting_started guide to reduce confusion
I'm a newcomer to Rails, I was reading the getting started guide today, and I carelessly mistook `update` with `edit`, which caused an error when running the app, and I worked really hard to debug but didn't find the cause. 

At last I realised It was my fault. I looked up on stackoverflow, and found other people made the same mistake as I did. Here is a link to the SO [post](https://stackoverflow.com/questions/27658767/rails-param-is-missing-or-the-value-is-empty-article). And happily I solved the problem.

I guess this can be a common mistake because `edit` and `update` share some similar meanings, they are all meant to modify something, so I guess we can do something to eliminate this confusion by:

1. point out in the very beginning that `edit` differs from `update` and
2.in the code example, we append the `edit` method, such that `edit` and `update` method exist at the same time, to let the readers know that they are different things.

What do you guys think?

English is not my first language, if words don't make sense, please ask me and give me a chance to explain to you.