unlinks Ruby on Rails Tutorial [skip ci]
I added myself links to the Ruby on Rails Tutorial years ago, after
meeting Michael in a RailsConf.

By then, the book was a commercial product, but had a free online
version and I thought it could be a good resource complementing the
official documentation.

Nowadays, only a few sample chapters are available for free, and I
don't consider these links to be fair with the rest of commercial
books about Rails anymore.