Maintain the consistency between `parts` and `value` of `AS::Duration`
In #37839 and #37849, we have modified to use `AS::Duration#value` instead of `parts` when `parts` is empty. Now, however, I think it's better to keep `parts` not be empty when `value` isn't zero.

Besides, the return value of `inspect` and `to_i` seems not to be consistent when the division had a remainder.

```ruby
(3.minutes / 2).inspect #=> "1 minute"
(3.minutes / 2).to_i #=> 90
```

To prevent these inconsistencies, we can rebuild the instance using `AS::Duration.build` when the division has a remainder.