Add regression test for url_helpers duplication
This adds a regression test that previously failed checking controllers
don't have multiple ancestors which include url_helpers or path_helpers.

This is checked by digging through the ancestors and seeing if they
include either the url_helpers_module or path_helpers_module and making
sure only one ancestor does that.

We usually don't test performance changes (and also would not prefer to
test implementation this closely), but because this has regressed in the
past, and I think it would be relatively easy to accidentally introduce
again, I think we should test this one.