Removed unnecessary words from Getting Started docs
Using "simple" and "just" in documentation is very tempting! It is there
to show enthusiasm and to encourage new people to try out the thing the
author loves. In this case, Rails. However, if you are following
documentation and struggle with it, reading these words can be a source
of frustration.

Also: using fewer words can make the instructions more clear and can
reduce ambiguity.