Prevent NoMethodError on named route in integration test.
After performing a request on a Metal controller the
controller instance on the integration session becomes
an instance of that Metal controller.

The next call to a named route that requires url_options
will throw a NoMethodError because a Metal controller
does not respond to url_options.

This changes the url_options method on the integration
session to check for the url_options method before
attempting to call it.