Use the 'What is FRAMEWORK?' title for all framework guides [ci skip]
Action View, Active Record and Active Storage use the 'What is
FRAMEWORK?' title for the framework introduction, the other frameworks
use 'Introduction'. The 'What is' form is more descriptive and better
for SEO.

This change makes the titles more consistent by using the 'What is
FRAMEWORK?' title for the other frameworks.