Move `name` key on configuration hash into `DatabaseConfig`
`name` is used by Rails to find the configuration by connection
specification name, but database adapters don't need to use `name` in
order to establish a connection. This is part of our work to separate
what the database needs to connect (the configuration hash) and the
what Rails needs to find connections (everything else).

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>